Ensured players would always be a unique color
Before if players connected and disconnected enough, they would eventually all default to the grey astronaut. Additionally, if the approach depended directly on the player count, there was a chance that multiple people would be the same main color (i.e., 1 red + 2 greens). This approach ensures everyone is a different non-grey color as long as there are 4 or less people