Implemented better method for syncing movements
The old method used NGO's built in transform and rigidbody components. These operated on an interpolated delay-based approach, which made input delay a big problem and made movement look weird in a lot of cases. This new method should fix both of those problems, but it introduces the problem of possible cheating, which I don't think is much of a concern for now, and rubber-banding, but that's handled in a way that shouldn't make it a big problem